requirements.txt
# Core Django
Django==5.1.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.1.0

# Database
dj-database-url==2.3.0
psycopg2-binary==2.9.10
psycopg==3.2.6
psycopg-binary==3.2.6

# Server
gunicorn==23.0.0

# Static files
whitenoise==6.9.0

# PDF Generation
fpdf==1.7.2
pillow==11.1.0
# weasyprint==65.1  # Temporarily disabled due to build issues

# Phone numbers
phonenumbers==9.0.5

# Environment variables
python-decouple==3.8
python-dotenv==1.1.0

# HTTP requests
requests==2.32.3

# Authentication
PyJWT==2.9.0

# SMS
africastalking==1.2.9

# Utilities
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.8.0
fonttools==4.58.0
idna==3.10
packaging==25.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
schema==0.7.7
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.0
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
zopfli==0.2.3.post1
